metadata: restore `metadata.json.in`
Move git revision placeholder directly into `metadata.json.in`.

Use meson's builtin `vcs_tag()`.

This should also remove extension bundle dependency on `package.json` -
which in turn should stop triggering tests for eslint updates.
