+++ .gitattributes
@@ -2,5 +2,5 @@
 #
 # SPDX-License-Identifier: CC0-1.0
 
-revision.txt.in export-subst
+metadata.json.in export-subst
 PKGBUILD export-subst
