+++ meson.build
@@ -97,42 +97,43 @@ gnome_shell = find_program(
   disabler: true,
 )
 
-git_revision_file = files('revision.txt.in')
+metadata = configuration_data()
+metadata.set_quoted('name', meson.project_name())
+metadata.set('version', meson.project_version())
+metadata.set_quoted('uuid', uuid)
+metadata.set_quoted('gettext_domain', gettext_domain)
+metadata.set_quoted('settings_schema', settings_schema)
+metadata.set('shell_versions', f'@shell_versions@'.replace('\'', '"'))
+
+metadata_configured = configure_file(
+  input: 'metadata.json.in',
+  output: 'metadata.json.in',
+  configuration: metadata,
+)
+
 git_revision_format = '%H'
 git_revision_placeholder = f'$Format:@git_revision_format@$'
 
-has_git_revision_placeholder = fs.read(git_revision_file).contains(
+has_git_revision_placeholder = fs.read('metadata.json.in').contains(
   git_revision_placeholder,
 )
 
 git = find_program('git', disabler: true, required: has_git_revision_placeholder)
 
 if has_git_revision_placeholder
-  git_revision_file = custom_target(
+  metadata_configured = vcs_tag(
     command: [
-      capture_stdout,
       git,
-      '-C',
-      '@CURRENT_SOURCE_DIR@',
-      '--git-dir',
-      '.git',
       'rev-list',
-      f'--format=format:@git_revision_format@',
       '-1',
+      f'--format=format:@git_revision_format@',
       '--no-commit-header',
       'HEAD',
     ],
-    output: 'revision.txt',
-    build_always_stale: true,
-    install: true,
-    install_dir: extension_dir,
-  )
-else
-  git_revision_file = fs.copyfile(
-    git_revision_file,
-    'revision.txt',
-    install: true,
-    install_dir: extension_dir,
+    input: metadata_configured,
+    output: 'metadata.json.vcs_tag',
+    replace_string: git_revision_placeholder,
+    fallback: 'unknown',
   )
 endif
 
@@ -140,31 +141,16 @@ subdir('tools')
 
 pack = [
   custom_target(
-    command: [
-      make_metadata_command,
-      '--package-json',
-      files('package.json'),
-      '--name',
-      meson.project_name(),
-      '--version-name',
-      meson.project_version(),
-      '--git-revision-file',
-      git_revision_file,
-      f'--uuid=@uuid@',
-      f'--gettext-domain=@gettext_domain@',
-      f'--settings-schema=@settings_schema@',
-      '--shell-version',
-      shell_versions,
-    ],
+    command: fix_metadata_command,
+    input: metadata_configured,
     output: 'metadata.json',
     install: true,
     install_dir: extension_dir,
   ),
-  # Both are "licenses" for autogenerated files (no copyright), should be fine
-  fs.copyfile('package-lock.json.license', 'metadata.json.license'),
+  fs.copyfile('metadata.json.in.license', 'metadata.json.license'),
   # Keep revision file in package to allow smooth upgrade (while old extension is loaded)
-  git_revision_file,
-  fs.copyfile('revision.txt.in.license', 'revision.txt.license'),
+  fs.copyfile('revision.txt', install: true, install_dir: extension_dir),
+  fs.copyfile('revision.txt.license'),
 ]
 
 foreach f : ['extension.js', 'prefs.js', 'AUTHORS']
