+++ metadata.json.in.license
@@ -1,0 +1,5 @@
+SPDX-FileCopyrightText: 2020 Aleksandr Mezin <mezin.alexander@gmail.com>
+SPDX-FileContributor: Pedro Sader Azevedo
+SPDX-FileContributor: Maximilien Richer
+
+SPDX-License-Identifier: CC0-1.0
