--- revision.txt.in
@@ -1,1 +1,0 @@
-$Format:%H$
