+++ tools/meson.build
@@ -11,9 +11,9 @@ make_zip_command = [
   '@INPUT@',
 ]
 
-make_metadata_command = [
-  find_program(files('makemetadata.py')),  # https://github.com/mesonbuild/meson/issues/11255
-  '--output',
+fix_metadata_command = [
+  find_program(files('fixmetadata.py')),  # https://github.com/mesonbuild/meson/issues/11255
+  '@INPUT@',
   '@OUTPUT@',
 ]
